fix: 미터 단위 표시 오류 수정

+++ src/main/kotlin/com/deepromeet/atcha/location/domain/POI.kt
@@ -7,9 +7,9 @@ data class POI(
     val location: Location,
     val businessCategory: String,
     val address: String,
-    val radius: Int? = null
+    val radius: Double? = null
 ) {
     fun distanceTo(other: Coordinate): POI {
-        return copy(radius = location.coordinate.distanceTo(other).toInt())
+        return copy(radius = location.coordinate.distanceTo(other))
     }
 }
